Refine Endpoint action for multi-cluster Service
When an Endpoint of a Multi-cluster Service is a local Service ClusterIP,
change the EndpointDNAT flow action to let it go to the corresponding exported
Service's group to select the final Endpoint. This can avoid that the traffic
goes out of the OVS bridge from antrea-gw0 (and handled by kube-proxy when it is
running) and comes back again.

Signed-off-by: Lan Luo <luola@vmware.com>